+++ src/Librarian.cpp
@@ -19,6 +19,6 @@ void Librarian::GetInfo()
 
 void Librarian::AddBook(Book* newBook) { library->GetBookSystem()->AddContent(newBook); }
 
-void Librarian::RemoveBook(int ISBN) { library->GetBookSystem()->RemoveContent(ISBN); }
+void Librarian::RemoveBook(long long ISBN) { library->GetBookSystem()->RemoveContent(ISBN); }
 
 void Librarian::SetLibrary(LibrarySystem* newLibrary) { library = newLibrary; }
